+++ http_handle_repo_tree.go
@@ -99,7 +99,7 @@ func httpHandleRepoTree(writer http.ResponseWriter, request *http.Request, param
 				files = append(files, displayTreeEntry{
 					Name:      string(name),
 					Mode:      fmt.Sprintf("%06o", mode),
-					Size:      int64(size),
+					Size:      size,
 					IsFile:    typeCode == 2,
 					IsSubtree: typeCode == 1,
 				})
